requirements.txt
pyephem==3.7.5.2
pyserial==2.7
